feat(cowfi): upgrade careers page
