+++ apps/cow-fi/app/careers/page.tsx
@@ -1,0 +1,22 @@
+'use server'
+
+import { getJobs } from '../../services/ashByHq'
+
+import { CareersPageContent } from '@/components/CareersPageContent'
+
+export default async function Page() {
+  const jobsData = (await getJobs()) || {}
+  const department = 'All'
+
+  const jobsCount = Object.keys(jobsData).reduce((acc, cur) => acc + jobsData[cur].length, 0)
+  const jobsCountForDepartment = department === 'All' ? jobsCount : jobsData[department].length
+
+  return (
+    <CareersPageContent
+      department={department}
+      jobsCountForDepartment={jobsCountForDepartment}
+      jobsCount={jobsCount}
+      jobsData={jobsData}
+    />
+  )
+}
