+++ apps/cow-fi/app/careers/refer-to-earn/layout.tsx
@@ -1,0 +1,9 @@
+import { Metadata } from 'next'
+
+export const metadata: Metadata = {
+  title: 'Refer-to-Earn',
+}
+
+export default function LayoutPage({ children }: { children: React.ReactNode }) {
+  return children
+}
